Change focus on double tap behavior
On tap, show a bottom bar with actions to start typing a query.
Add two buttons: one to edit the current query and one to start fresh.
Space is left for a 3rd button for additional settings, if needed.

The bar is hidden as soon as possible and should not be a hidrance.

This is similar to the previous behavior but less obscure. It is still
possible to quickly type search by double tapping at the right place.